+++ src/main/java/org/jabref/JabRefMain.java
@@ -11,6 +11,7 @@ import org.jabref.cli.ArgumentProcessor;
 import org.jabref.cli.JabRefCLI;
 import org.jabref.gui.FXDialog;
 import org.jabref.gui.remote.JabRefMessageHandler;
+import org.jabref.logic.exporter.ExporterFactory;
 import org.jabref.logic.journals.JournalAbbreviationLoader;
 import org.jabref.logic.l10n.Localization;
 import org.jabref.logic.net.ProxyAuthenticator;
@@ -168,7 +169,11 @@ public class JabRefMain extends Application {
                 Globals.prefs.getXMPPreferences(), Globals.getFileUpdateMonitor());
         Globals.entryTypesManager.addCustomOrModifiedTypes(preferences.loadBibEntryTypes(BibDatabaseMode.BIBTEX),
                 preferences.loadBibEntryTypes(BibDatabaseMode.BIBLATEX));
-        Globals.exportFactory = Globals.prefs.getExporterFactory(Globals.journalAbbreviationRepository);
+        Globals.exportFactory = ExporterFactory.create(
+                Globals.prefs.getCustomExportFormats(Globals.journalAbbreviationRepository),
+                Globals.prefs.getLayoutFormatterPreferences(Globals.journalAbbreviationRepository),
+                Globals.prefs.getSavePreferencesForExport(),
+                Globals.prefs.getXMPPreferences());
 
         // Initialize protected terms loader
         Globals.protectedTermsLoader = new ProtectedTermsLoader(Globals.prefs.getProtectedTermsPreferences());